+++ packages/ui/VirtualList/VirtualListBasic.js
@@ -347,6 +347,17 @@ class VirtualListBasic extends Component {
 	}
 
 	componentDidUpdate (prevProps, prevState) {
+		const items = document.getElementsByClassName(css.listItem);
+		if (!this.itemMarginTop && this.itemMarginTop !== 0 && items.length > 0) {
+			const firstItemStyle = window.getComputedStyle(items[0].children[0]);
+			this.itemMarginTop = Number(firstItemStyle.getPropertyValue('margin-top').slice(0, -2));
+			this.itemMarginBottom = Number(firstItemStyle.getPropertyValue('margin-bottom').slice(0, -2));
+			this.itemMarginLeft = Number(firstItemStyle.getPropertyValue('margin-left').slice(0, -2));
+			this.itemMarginRight = Number(firstItemStyle.getPropertyValue('margin-right').slice(0, -2));
+		}
+		this.scrollBounds.maxTop += this.itemMarginTop + this.itemMarginBottom;
+		this.scrollBounds.maxLeft += this.itemMarginLeft + this.itemMarginRight;
+
 		let deferScrollTo = false;
 		const {firstIndex, numOfItems} = this.state;
 
@@ -409,13 +420,39 @@ class VirtualListBasic extends Component {
 			!equals(prevProps.itemSize, this.props.itemSize) ||
 			(!this.hasDataSizeChanged && !shallowEqual(prevProps.itemSizes, this.props.itemSizes))
 		) {
+			const {clientHeight, clientWidth, scrollHeight, scrollWidth} = this.scrollBounds;
+
+			/* istanbul ignore next */
+			if (this.scrollPosition === 0 && this.moreInfo.firstVisibleIndex !== 0) {
+				this.scrollPosition = this.prevScrollPosition;
+				this.setState({firstIndex: prevState.firstIndex});
+				this.updateMoreInfo(this.props.dataSize, this.scrollPosition);
+
+				for (let i = prevState.firstIndex; i < prevState.firstIndex + this.state.numOfItems; i++) {
+					const size = this.isPrimaryDirectionVertical ? clientHeight : clientWidth;
+
+					if (this.itemPositions[i + 1]?.position > this.prevItemPositions[i + 1]?.position) {
+						if (i === this.moreInfo.lastVisibleIndex && this.getItemBottomPosition(i) > this.scrollPosition + size) {
+							this.scrollPosition = this.getItemBottomPosition(i) - size;
+						}
+						break;
+					}
+
+					if (this.itemPositions[i + 1]?.position < this.prevItemPositions[i + 1]?.position) {
+						if (this.itemPositions[i].position < this.scrollPosition) {
+							this.scrollPosition = this.itemPositions[i].position;
+						}
+						break;
+					}
+				}
+			}
+
 			const {x, y} = this.getXY(this.scrollPosition, 0);
 
 			this.calculateMetrics(this.props);
-			this.setState(this.getStatesAndUpdateBounds(this.props));
+			this.setState(this.getStatesAndUpdateBounds(this.props, prevState.firstIndex ? prevState.firstIndex : 0));
 			this.setContainerSize();
 
-			const {clientHeight, clientWidth, scrollHeight, scrollWidth} = this.scrollBounds;
 			const xMax = scrollWidth - clientWidth;
 			const yMax = scrollHeight - clientHeight;
 
@@ -471,17 +508,23 @@ class VirtualListBasic extends Component {
 	shouldUpdateBounds = false;
 
 	dimensionToExtent = 0;
+	itemMarginLeft = null;
+	itemMarginRight = null;
+	itemMarginTop = null;
+	itemMarginBottom = null;
 	threshold = 0;
 	maxFirstIndex = 0;
 	curDataSize = 0;
 	hasDataSizeChanged = false;
 	cc = [];
 	scrollPosition = 0;
+	prevScrollPosition = 0;
 	scrollPositionTarget = 0;
 	scrollToPositionTarget = -1;
 
 	// For individually sized item
 	itemPositions = [];
+	prevItemPositions = [];
 	indexToScrollIntoView = -1;
 
 	updateScrollPosition = ({x, y}, behavior) => {
@@ -556,15 +599,16 @@ class VirtualListBasic extends Component {
 		const maxPos = isPrimaryDirectionVertical ? scrollBounds.maxTop : scrollBounds.maxLeft;
 		const position = this.getGridPosition(index);
 		let offset = 0;
+		const marginOffset = isPrimaryDirectionVertical ? this.itemMarginTop + this.itemMarginBottom : this.itemMarginLeft + this.itemMarginRight;
 
 		if (stickTo === 'start') {         // 'start'
 			offset = optionalOffset;
 		} else if (this.props.itemSizes) { // 'end' for different item sizes
-			offset = primary.clientSize - this.props.itemSizes[index] - optionalOffset;
+			offset = primary.clientSize - this.props.itemSizes[index] - marginOffset - optionalOffset;
 		} else if (stickTo === 'center') { // 'center'
 			offset = (primary.clientSize / 2) - (primary.gridSize / 2) - optionalOffset;
 		} else {                           // 'end' for same item sizes
-			offset = primary.clientSize - primary.itemSize - optionalOffset;
+			offset = primary.clientSize - primary.itemSize - marginOffset - optionalOffset;
 		}
 
 		/* istanbul ignore next */
@@ -655,6 +699,10 @@ class VirtualListBasic extends Component {
 		this.secondary = secondary;
 
 		// reset
+		this.prevScrollPosition = this.scrollPosition;
+		for (let i = 0; i < this.itemPositions.length; i++) {
+			this.prevItemPositions[i] = this.itemPositions[i];
+		}
 		this.scrollPosition = 0;
 		this.scrollToPositionTarget = -1;
 		if (scrollMode === 'translate' && this.contentRef.current) {