+++ packages/sampler/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 The following is a curated list of changes in the Enact Sampler, newest changes on the top.
 
+## [5.4.1] - 2025-12-30
+
+No significant changes.
+
+## [5.4.0] - 2025-12-08
+
+No significant changes.
+
+## [5.3.2] - 2025-10-28
+
+No significant changes.
+
+## [5.3.1] - 2025-10-14
+
+No significant changes.
+
 ## [5.3.0] - 2025-09-24
 
 No significant changes.