+++ packages/spotlight/CHANGELOG.md
@@ -2,11 +2,21 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
-## [unreleased]
+## [5.4.1] - 2025-12-30
+
+No significant changes.
+
+## [5.4.0] - 2025-12-08
 
 ### Fixed
 
-- `spotlight` navigation from an element to a different container on page load.
+- `spotlight` to skip visibility check when entering a restricted container or choose a visible container as next target
+
+## [5.3.2] - 2025-10-28
+
+### Fixed
+
+- `spotlight` navigation from an element to a different container on page load
 
 ## [5.3.1] - 2025-10-14