+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,31 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Changed
 
 - `ui/Scroller` scroll animation method for `scrollMode: native`
+### Fixed
+
+- `ui/VirtulList` not to cut off focused item when scroll by pressing key
+- `ui/VirtualList` not to jump to the top when item size was changed
+- `ui/VirtualList.VirtualGridList` to adjust itemSize properly when resizing window 
+
+## [5.4.1] - 2025-12-30
+
+No significant changes.
+
+## [5.4.0] - 2025-12-08
+
+### Added
+
+- `limestone/Card` `splitCaption` prop to allow user to split the content for the captions
+- `ui/useScroll.ScrollAnimator` `oase-out-cubic` animation function
+
+### Changed
+
+- `ui/useScroll.ScrollAnimator` default animation function
+
+## [5.3.2] - 2025-10-28
+
+No significant changes.
+
 ## [5.3.1] - 2025-10-14
 
 ### Fixed